--- a/app/sftp/ftp_client.py
+++ b/app/sftp/ftp_client.py
@@ -61,6 +61,7 @@ def upload_zip(sftp, zip_data, filename, statsd_client):
             ))
 
     with sftp.open('{}/{}'.format(sftp.pwd, filename), mode='w') as remote_file:
+        remote_file.set_pipelined()
         remote_file.write(zip_data)
 
     statsd_client.timing("ftp-client.zip-upload-time", monotonic() - start_time)