@sakisv 2021-01-18
Testing on staging:

With sftp:
```
100MB_file.txt                         100%   99MB   3.0MB/s   00:32
```

With python:

```
uploading zip 100MB_file.txt of total size 104,100,000
Data 100MB_file.txt uploaded to DVLA

real    0m35.413s
user    0m2.638s
sys     0m0.392s
```